Split finetune annotation services
